reconsturct all records

+++ jsonstore/models.py
@@ -4,10 +4,16 @@ from django.db import models
 from django.dispatch import receiver
 from django.db.models.signals import pre_save, post_save
 from django.contrib.postgres.fields import JSONField
-from mptt.models import MPTTModel, TreeForeignKey
+#from mptt.models import MPTTModel, TreeForeignKey
 from urllib import error, request
 import sys, random, time, json
-from moac.models import * 
+from decimal import Decimal
+from moac.models import Ledger, Address, Transaction, Uncle
+
+#class JsonStat(models.Model):
+#	metric = models.CharField(max_length=16,unique=True)
+#	data = JSONField()
+#	date = models.DateField(auto_now=True)
 
 class JsonJingtumLedger(models.Model):
 	hash = models.CharField(max_length=64,unique=True,editable=False)
@@ -69,29 +75,30 @@ class JsonMoacLedger(models.Model):
 			return True
 		if not url:
 			url = "http://localhost:3003/api/uncle/"
-		for index in range(len(self.data['uncles'])):
-			sys.stdout.write("\t...trying %s/%s\n" % (self.id,index))
-			try:
-				response = request.urlopen("%s%s/%s" % (url,self.id,index), timeout=30)
-				if response.status == 200:
-					result = json.loads(response.read().decode())
-					hash = result['hash']
-					uncle = JsonMoacUncle(hash=hash,ledger=self,data=result)
-					uncle.save()
-					out = sys.stdout.write("... retrieved uncle for %s/%s\n" % (self.id,index))
-				else:
-					out = sys.stdout.write("..!..http returned status %s\n" % response.status)
+		if self.data['uncles']:
+			for index in range(len(self.data['uncles'])):
+				sys.stdout.write("\t...trying %s/%s\n" % (self.id,index))
+				try:
+					response = request.urlopen("%s%s/%s" % (url,self.id,index), timeout=30)
+					if response.status == 200:
+						result = json.loads(response.read().decode())
+						hash = result['hash']
+						uncle = JsonMoacUncle(hash=hash,ledger=self,data=result)
+						uncle.save()
+						out = sys.stdout.write("... retrieved uncle for %s/%s\n" % (self.id,index))
+					else:
+						out = sys.stdout.write("..!..http returned status %s\n" % response.status)
+						return False
+				except Exception as e:
+					out = sys.stderr.write("... exception happend for %s/%s\n" % (self.id,index))
+					print(e)
 					return False
-			except Exception as e:
-				out = sys.stderr.write("... exception happend for %s/%s\n" % (self.id,index))
-				print(e)
-				return False
 		self.synced = True
 		self.save()
-		sys.stdout.write("\tsynced unclues for ledger %s\n" % self.id)
+		sys.stdout.write("\t...synced uncles\n")
 		return True
 
-	def proc_ledger(self,do_uncle=False):
+	def proc_ledger(self,do_uncle=False, bypass_balance=False):
 		addresses = set()
 		try:
 			ledger = Ledger.objects.get(hash=self.hash)
@@ -101,43 +108,35 @@ class JsonMoacLedger(models.Model):
 					jmu = JsonMoacUncle.objects.get(hash=uncle_hash,ledger=JsonMoacLedger.objects.get(hash=ledger.hash))
 					miner,created = Address.objects.get_or_create(address=jmu.data['miner'])
 					addresses.add(miner)
-					if created:
-						miner.update_display()
 					uncle.miner = miner
 					uncle.number = jmu.data['number']
 					uncle.save()
 		except Ledger.DoesNotExist:
 			miner,created = Address.objects.get_or_create(address=self.data['miner'])
 			addresses.add(miner)
-			if created:
-				miner.update_display()
 			timestamp=self.data['timestamp']
 			if self.id == 0:
-				timestamp = JsonMoacLedger.objects.get(id=1).data['timestamp'] - 10
-				ledger = Ledger(hash=self.hash, number=self.id, difficulty = self.data['difficulty'], nonce = self.data['nonce'], miner=miner, timestamp=timestamp)
+				timestamp = 1525064660
+				ledger = Ledger(hash=self.hash, id=self.id, difficulty = self.data['difficulty'], nonce = self.data['nonce'], miner=miner, timestamp=timestamp)
 				ledger.save()
 			else:
-				duration = int(timestamp - Ledger.objects.get(number=self.id -1).timestamp)
+				duration = int(timestamp - Ledger.objects.get(id=self.id -1).timestamp)
 				num_txs = len(self.data['transactions'])
 				tps = int(num_txs / duration)
-				ledger = Ledger(hash=self.hash, number=self.id, num_txs=num_txs, tps=tps, duration=duration, difficulty = self.data['difficulty'], nonce = self.data['nonce'], miner=miner, timestamp=timestamp)
+				ledger = Ledger(hash=self.hash, id=self.id, num_txs=num_txs, tps=tps, duration=duration, difficulty = self.data['difficulty'], nonce = self.data['nonce'], miner=miner, timestamp=timestamp)
 				ledger.save()
 				if self.data['transactions']:
-					sys.stdout.write('update transactions:\n')
+					sys.stdout.write('update transactions:\n\t...')
 					for txr in self.data['transactions']:
 						sys.stdout.write("%s, " % txr['transactionIndex'])
 						tx_from, created = Address.objects.get_or_create(address=txr['from'])
 						addresses.add(tx_from)
-						if created:
-							tx_from.update_display()
 						if txr['to']:
 							tx_to, created = Address.objects.get_or_create(address=txr['to'])
 							addresses.add(tx_to)
-							if created:
-								tx_to.update_display()
 						else:
 							tx_to = None
-						transaction, created = Transaction.objects.get_or_create(ledger=ledger,hash=txr['hash'],tx_from=tx_from, tx_to=tx_to, value=int(float(txr['value'])) / 1000000000, index=int(txr['transactionIndex']))
+						transaction, created = Transaction.objects.get_or_create(ledger=ledger,hash=txr['hash'],tx_from=tx_from, nonce=txr['nonce'], tx_to=tx_to, value=int(float(txr['value'])) / 1000000000, index=int(txr['transactionIndex']))
 						transaction.save()
 					sys.stdout.write('\n')
 				for uncle_hash in self.data['uncles']:
@@ -145,16 +144,16 @@ class JsonMoacLedger(models.Model):
 					jmu = JsonMoacUncle.objects.get(hash=uncle_hash,ledger=JsonMoacLedger.objects.get(hash=ledger.hash))
 					miner,created = Address.objects.get_or_create(address=jmu.data['miner'])
 					addresses.add(miner)
-					if created:
-						miner.update_display()
 					uncle.miner = miner
 					uncle.number = jmu.data['number']
 					uncle.save()
-			sys.stdout.write('update balances:\n')
-			for address in addresses:
-				sys.stdout.write("%s, " % address.address)
-				address.update_balance()
-			sys.stdout.write('\n')
+			if not bypass_balance:
+				sys.stdout.write('update balances:\n\t...')
+				for address in list(filter(lambda x: not x.is_contract and not x.display.startswith('system') , addresses)):
+					a = Address.objects.get(address=address.address)
+					sys.stdout.write("%s, " % a.display)
+					a.update_balance()
+				sys.stdout.write('\n')
 			
 	@classmethod
 	def verify(cls,start=0):
@@ -180,15 +179,19 @@ class JsonMoacLedger(models.Model):
 		if not url:
 			url = "http://localhost:3003/api/block/%s" % height
 		done = False
+		last = None
+		ledger = None
 		try:
-			last = cls.objects.get(id=height-1)
+			if height > 0:
+				last = cls.objects.get(id=height-1)
 			ledger = cls.objects.get(id=height)
 			done = True
-			if ledger.parent_hash != last.hash:
-				sys.stdout.write("\tinconsistancy found, delete last two")
+			if last and ledger.parent_hash != last.hash:
+				sys.stdout.write("\tinconsistancy found, delete last two\n")
 				last.delete()
 				ledger.delete()
-				ledger = cls.objects.get(id=height-2)
+				if height > 1:
+					ledger = cls.objects.get(id=height-2)
 		except cls.DoesNotExist:
 			pass
 		while not done:
@@ -198,13 +201,18 @@ class JsonMoacLedger(models.Model):
 					result = json.loads(response.read().decode())
 					hash = result['hash']
 					parent_hash = result['parentHash']
-					if parent_hash == last.hash:
+					if height > 0 and last and parent_hash == last.hash:
+						ledger = cls(id=height,hash=hash,parent_hash=parent_hash,data=result)
+						ledger.save()
+					elif height == 0:
 						ledger = cls(id=height,hash=hash,parent_hash=parent_hash,data=result)
 						ledger.save()
 					else:
-						sys.stdout.write("\tinconsistancy found, delete last two")
-						last.delete()
-						ledger = cls.objects.get(id=height-2)
+						sys.stdout.write("\tinconsistancy found, delete last two\n")
+						if last:
+							last.delete()
+						if height > 1:
+							ledger = cls.objects.get(id=height-2)
 					done = True
 				else:
 					out = sys.stdout.write("..!..http returned status %s\n" % response.status)
@@ -247,4 +255,6 @@ def post_save_ledger_moac(sender, instance, created, **kwargs):
 		# retrieve JsonMoacUncles
 		instance.sync_uncles()
 		# generate ledgers, transactions and uncles
+		#instance.proc_ledger(bypass_balance=True)
 		instance.proc_ledger()
+
